requirements.txt
flask==3.1.3
flask-restx==1.3.0
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
marshmallow==3.23.2
python-dotenv==1.0.1
requests>=2.31.0
gunicorn>=22.0.0
pytest==8.3.4
pytest-cov==6.0.0
Werkzeug==3.1.3